Add support for overrideConsoleMethods & printToConsole
